chore: refactor a bunch
