+++ src/index.js
@@ -1,0 +1,8 @@
+import get from './get.js';
+import util from './util.js';
+
+const { findpath } = util;
+
+export {
+    findpath,
+}
