+++ src/util.js
@@ -1,4 +1,9 @@
 import fs from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
+import url from 'node:url';
+
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 /**
  * Check if file exists at specified path.
@@ -34,26 +39,51 @@ const EXE_NAME = {
     linux: "nw",
 };
 
-/** 
- * Get the compressed file name.
+/**
+ * Get the platform dependant path of the NW.js or ChromeDriver binary.
  * 
- * @param {string} version
- * @param {string} flavor
- * @param {string} platform
- * @param {string} arch
- * @returns {string}
+ * @param {'nwjs' | 'chromedriver'} executable Path to NW.js or Chromedriver executable.
+ * @return {string}
  */
-function getCompressedFile(version, flavor, platform, arch) {
-    return [
-        'nwjs',
-        flavor === 'normal' ? '' : `-${flavor}`,
-        '-v',
-        version,
-        '-',
-        PLATFORM_KV[platform],
-        '-',
-        ARCH_KV[arch],
-    ].join('');
+function findpath(executable = 'nwjs') {
+    const nwDir = path.resolve(__dirname, '..', 'nwjs');
+    /**
+     * File path to executable.
+     * 
+     * @type {string}
+     */
+    let binPath = '';
+
+    /**
+      * Host operating system
+      * 
+      * @type {NodeJS.Platform}
+      */
+    let hostOs = process.env.npm_config_nwjs_platform || process.env.NWJS_PLATFORM || process.platform;
+
+    /**
+     * Get the platform dependant path of the NW.js binary.
+     */
+    function findNwjs() {
+        binPath = path.resolve(nwDir, EXE_NAME[hostOs]);
+    }
+
+    /**
+     * Get the platform dependant path of the ChromeDriver binary.
+     */
+    function findChromeDriver() {
+        binPath = path.resolve(nwDir, `chromedriver${process.platform === "win32" ? ".exe" : ""}`);
+    }
+
+    if (executable === 'nwjs') {
+        findNwjs();
+    } else if (executable === 'chromedriver') {
+        findChromeDriver();
+    } else {
+        console.error(`[ ERROR ] Expected nwjs or chromedriver, got ${executable}.`);
+    }
+
+    return binPath;
 }
 
-export default { ARCH_KV, EXE_NAME, PLATFORM_KV, fileExists, getCompressedFile }
+export default { ARCH_KV, EXE_NAME, PLATFORM_KV, fileExists, findpath }
